Add Capistrano task to upload gulp build files
Add Capistrano task to upload gulp build files

We have a `src/Frontend/Themes/Triskel/src` folder.
`gulp build` builds this to `src/Frontend/Themes/Triskel/Core`.
This `Core` folder is not in our repo,
so when deploying we need to upload those files.
This is exactly what this new Capistrano task does.